Patrick/David Reorganize serializers into separate package
