+++ src/main/java/org/mercycorps/translationcards/txcmaker/serializer/CardSerializer.java
@@ -1,9 +1,10 @@
-package org.mercycorps.translationcards.txcmaker.model;
+package org.mercycorps.translationcards.txcmaker.serializer;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonSerializationContext;
 import com.google.gson.JsonSerializer;
+import org.mercycorps.translationcards.txcmaker.model.Card;
 
 import java.lang.reflect.Type;
