+++ src/main/java/org/mercycorps/translationcards/txcmaker/serializer/FinalizedLanguageSerializer.java
@@ -1,4 +1,4 @@
-package org.mercycorps.translationcards.txcmaker.wrapper;
+package org.mercycorps.translationcards.txcmaker.serializer;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
