+++ src/main/java/org/mercycorps/translationcards/txcmaker/serializer/GsonWrapper.java
@@ -1,4 +1,4 @@
-package org.mercycorps.translationcards.txcmaker.wrapper;
+package org.mercycorps.translationcards.txcmaker.serializer;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
