+++ src/main/java/org/mercycorps/translationcards/txcmaker/service/StorageService.java
@@ -3,7 +3,7 @@ package org.mercycorps.translationcards.txcmaker.service;
 import com.google.api.client.util.IOUtils;
 import org.mercycorps.translationcards.txcmaker.model.deck.Deck;
 import org.mercycorps.translationcards.txcmaker.model.deck.DeckMetadata;
-import org.mercycorps.translationcards.txcmaker.wrapper.GsonWrapper;
+import org.mercycorps.translationcards.txcmaker.serializer.GsonWrapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
