+++ src/test/java/org/mercycorps/translationcards/txcmaker/serializer/FinalizedDeckSerializerTest.java
@@ -1,4 +1,4 @@
-package org.mercycorps.translationcards.txcmaker.wrapper;
+package org.mercycorps.translationcards.txcmaker.serializer;
 
 import com.google.gson.JsonObject;
 import com.google.gson.JsonSerializationContext;
@@ -6,6 +6,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mercycorps.translationcards.txcmaker.model.FinalizedDeck;
 import org.mercycorps.translationcards.txcmaker.model.FinalizedLanguage;
+import org.mercycorps.translationcards.txcmaker.serializer.FinalizedDeckSerializer;
 
 import java.util.ArrayList;
