+++ src/test/java/org/mercycorps/translationcards/txcmaker/serializer/GsonWrapperTest.java
@@ -1,4 +1,4 @@
-package org.mercycorps.translationcards.txcmaker.wrapper;
+package org.mercycorps.translationcards.txcmaker.serializer;
 
 import org.junit.Test;
 import org.mercycorps.translationcards.txcmaker.model.Card;
